requirements.txt
# ocr_requirements.txt - Minimal requirements for OCR functionality
pytesseract>=0.3.8
pdf2image>=1.16.0
numpy>=1.20.0
opencv-python>=4.5.1
Pillow>=8.2.0